Added UART and fixed PORTMUX
Added conditional groups for PORTMUX, checking if register exist for selected MCU. This should work for all ATtiny series so the core stays the same with minor changes.
Might have to add alternative PORTMUX for the ATtiny series, using default for now.
Fixed the UART and Serial in Arduino is working now. Able to receive messages with UART to USB converter.