PORTMUX for xy2
Fix #46 - should work for UART. Unsure if it will work for TWI. The
datasheet implies that it's not possible to use TWI on the xy2 series,
because only the alternate pins are exposed, but the portmux register
doesn't show an option to put TWI on the alternate pins. If the
datasheet is wrong and that portmux option is available, just not
listed, then this should do the trick. Otherwise, there's no TWI on the
xy2, and it's not our fault.