Update MakeUPDIProgrammer.md
Adjust pin order to reflect @freemovers concern with the connector being connected backwards - which is a very good point, might I add, as people (including myself) are forever connecting the ISP connector on classic AVRs backwards. Upon reflection, that the AVR ISP connector doesn't trash the board when connected backwards is probably not an accident - if any pin other than reset ended up connected to Vcc when the connector was backwards, this would not be the case. .