Don't specify the upload tool versions in platform.txt
This change will cause the tool versions specified in the toolsDependencies field of the Boards Manager JSON file to be used with Boards Manager installations. Manual installations will use the newest installed version of the tools. The reason for making this change is it eliminates the need to specify the tool version in two separate locations: platform.txt before the release and package_drazzy.com_index.json after the release, reducing the maintenance burden and the likelihood of ending up with a mismatch.

The reason the tools versions were originally specified in platform.txt is because those lines were copied intact from Arduino megaAVR Boards. I believe the reason Arduino did that is because Arduino Web Editor does not match the Arduino IDE's behavior of determining tool versions by looking at the Boards Manager JSON file, so it ends up using a random tool version if it's not specified in platform.txt.