Add Xplained416 and explained817 board types that always use bootloader (since programming via xplainedmini_updi doesn't seem to work.)
Set up the rest of the bootloader menus for xy7, xy6, xy4, xy2.
Add upload.speed to the upload pattern.
Change the bootloader.pattern so that it only changes the BOOTEND fuse.  So if you want special fuse values, set them up for the board and use "upload using programmer" (with an appropriate programmer), and THEN "burn boorloader" (I don't know if this works yet.)