constify F_CPU_CORRECTED
The F_CPU_CORRECTED is currently always equal to F_CPU and cannot
change. Since it is variable, however, the compiler cannot optimize the
calculations using it (like those calculating baud rate in
Tone/Wire/UART) to compile time and the code that is produced is larger.

I've done very simple tests on t416 with:
1. Empty sketch (just empty setup() and loop()) and the sketch was
   reduced from 636 bytes to 586 bytes.
2. An empty sketch with simple tone(4, 1000) call in the setup and
   nothing more and the sketch size was reduced from 1200 bytes to 1016
   bytes.
3. An empty sketch with just a "Wire.h" include and nothing more and the
   sketch size was reduced from 2926 bytes to 2740 bytes.

Not an impressive result but the change is also quite straightforward
and should be safe.

Related to #54.