twi.h; pack enums
By default enums in C are ints and int on AVR8 is 16-bit making our
enums also 16bits even though they will easily fit in 8-bits. This means
that we need to use more instructions when operating on them which makes
code less compact and less efficient.

This could be globally changed with --short-enums gcc flag but it can be
dangerous if somebody is using some object files with enums compiled
without it. To make change less intrusive, we can change the size of
selected enums with packed attribute.

This reduces the size of an empty sketch with just Wire.h imported from
2926 to 2852 bytes in my case (74 bytes).