twi: operate on khz
Operation on 32-bit values takes a lot of instructions on our 8bit CPUs.
In this case this can be avoided as we know we are only operating on khz
values anyways so we can pre-compute it. The value in khz fits nicely in
16-bit so the generated code will be more compact. As an added bonus I
think it also makes those values easier to read (you don't have to count
those zeros :)).

This reduced size of my sketch from 2926 to 2828 (98 bytes) on my t416.

This is related to #54.