Move TCA0 timekeeping to HUNF interrupt, fix bug with TCD0 as default timer
Using HUNF is slightly preferable to LUNF because on 8-pin parts, 3 of
the 4 pwm pins are controlled by the low byte registers; this allows
them to be set to run at a higher frequency without disrupting millis
timekeeping even when TCA0 is used for millis timekeeping. With LUNF
used as the interrupt, only the single pin on the high byte could have
it's PWM frequency increased in that way.

Also fix bug where TCD0 timekeeping ran at double-speed and PWM duty
cycles were wrong when default millis timer was selected and that was
TCD0.