Correct #189: micros() glitches with TCA0 as millis source in 1.8.9 and later
1.8.9 enhanced the configuration of TCA0 when used as source for millis
by using the high byte instead of the low byte. This would provide no
change in functionality on most parts, but would permit the user to
change the frequency on the three pins controlled by the low half of the
timer without impacting millis, instead of the single pin controlled by
the high half of the timer (on all other parts, there are three on each
half). As originally introduced, several references to the low timer
remained in the millis code, and this resulted in timing glitches when
using micros() - this changes those cases to correctly point to the high
timer.