Add CI workflow to check for misspelled words
On every push and pull request, the files of the repository will be checked for commonly misspelled words using codespell.

The workflow is also run weekly in order to catch new misspelled word detections that result from updates to the codespell dictionary.

In the event of a false positive, add the word in all lower case to megaavr/extras/ci/codespell-ignore-words-list.txt