Add a CI workflow to report memory usage change caused by PRs
On every push, change in memory usage is calculated and displayed in the workflow build log.
A JSON format file containing this information as well as other data on the compilations is saved as a workflow log.

The memory usage change report will be commented to the pull request thread.