Fix Servo library!
This fixes #241 and #195.

There were a LOT of things broken about this library...

Turned up a way in which wiring.c may violate expectations of libraries (in fact, it *did* break pulseIn(), just nobody had noticed) when TCD0 was used as millis source, now there's separate macros to convert between microseconds and millisClockCycle and clockCycle. Completed conversion of code to work without the TCA prescaler (this was where most of the problems came from, though not all of them.
presence or absence of used attribute on _timer in ServoTimers has unpredictable (though small) impact on size of compiled binary... but putting it there gets rid of a scary warning; so is probably worth it even if it costs half a dozen bytes on average. The actual generated code though, has differences all over the place - it causes a bunch of different optimization decisions to be made, it would appear.