Improve digital I/O error checking, digitalWrite/pinMode emulate classic AVRs
Improvements to error checking and NOT_A_PIN exception should fix #278

digitalWrite() and pinMode() will now emulate classic AVR behavior with regards to the state of the port output register when digitalWrite() is called on an INPUT pin. Setting pinMode() to INPUT or INPUT_PULLUP will also set the port output register in the same way. Should fix #279 - and god knows how many other issues not yet reported; the `digitalWrite(pin,HIGH);` prior to `pinMode(pin,OUTPUT);` trick is very widespread in the wild, and breaking compatibility with it is foolish. I am honestly shocked that this is the first I've heard of it.