Un-mess-up tinyNeoPixel libraries
The copies on DxCore and here had diverged significantly. Also, the way that the presence/absence of a valid pin was handled here was nonsensical... t was checled for.... in places where it did not matter. But in the one place where it did, show, the same check was not performed. The null pointer would be treated as the pointer to the port struct of the pin (port struct address + 4 gives PORTx.OUT -so null pointer + 4 gives 0x0004. That's VPORTB.DIR... And because the bitmask was coming back NOT_A_PIN, in the event that it wound up getting an invalid pin iot would merrily toggle the hell out of the direction register for PORTB. Also, in some places the pin was an int8_t, in others uint8_t. Now is uint8_t everywhere in the library, consistent with conventions elsewhere. If pin is set to something that couldn't be a digital pin number. we now refuse to output signals.