Critical Bugfix
fix bug with constrain and round macros.

A long-standing bug woth significant performance and flash usage ramifications was discovered in analogWrite and turnOffPWM functions wherein  a loop could be run during the course of turning pwm on or off, in the case of TCD, potentially while interrupts were disabled. It could execute for as many as 7 iterations in the worst case, at 4 clocks each, Additional efficiencies were also found in turnOffPWM().