Actually fix critical bootloader bug #320
Okay, I was now able to fix it - it was a misplaced pair of quotes causing an argument to be treated wrong. Still, am hoping to get an updated version of the programming tool for a future release which will error instead of silently failing and producing a half-written device if it gets bad arguments passed to it.