New version of pymcuprog bridge
Now passing arguments to it which it doesn't understand results in an error, not silent failure with result not matching user expectations.

Also critical fix to pymcuprog - with verbose debugging on. it would pass certain things read from the chip to the console without sanitizing them, One of the fields it passed like that contained a \0 (null). This would cause output to be truncated at the naked null in many programs (anything written in a language that used null-terminated strings) including the Arduino IDE.