spelling, fix a warning.
If resolution was not compile-time known, old code would generate (spurious, as far as I can tell) warnings about sampnum potentially being used uninitialized despite the fact that sampnum was initialized if res > 0x80. sampnum was only used later if res > 0x80 (else it was not used), that seemed to be enough to confuse the compiler into emitting a warning.