Correct issue with TCB0 as millis breaking compile due to tone.cpp
Fixes #394. Even when tone wasn't used, it would generate compile errors that stopped the sketch from buiilding (this was not from an intentional attempt to stop compile - not sure why it didn't always do this,! But we just checked if millis was TCB0, if it was, assumed we have a TCB1, and then get undefined reference errors when we tried to refer to TCB1.