Whoever wrote this original code should be ashamed (#400)
This was code from a case where I was told that the official core fixed this bug already, There's a huge problem with wire, you need to get the fix in before the next release (which I was in the process of releasing) with Wire. Okay, well, the official core is going to work without a hitch, I guess I'll go grab it. And there was a problem. what I wound up getting wasn't complete and was even worse, so I had to do another emergency release to fix it.....

Having had a deep dive into how the fuck you're supposed to calculate the baud value to use, It looks to me like they didn't even TRY or THINK at any point in the process of writing that code. And, well, when you don't try to do it right and do a bunch of stuff in silly and/or naïve ways the result, you get crap.

Can you test this code out and see if it actually works? I don't want to do another release that shits on Wire,  but I also don't have a dedicated wire test rig - I should have one for tinyAVR and Dx both... But  I should have a lot of things that I won't have time to build any time soon.