ensure consecutive calls to analogWrite() work on TCD pins
#414
The previous check was only testing if it was *EITHER* enabled, or available to accept a command. We need to wait for it to be both,. I think this happened when I switched to using SYNCEOC instead of SYNC because people complained about the glitch that could occur when you changed a PWM durty cycle and missed set the compare to a value that had already been passed, resulting in the pin statying in the wrong state for the rest of the cycle. SYNCEOC fixes that by doing it at the end of a cycle, but takes, well, a pwm cycle. Which is less time that it takes for the next analogWrite to come along and want to issue it's own synceoc. And my check tested for either status bit instead of both status bitrs.