Update some examples
RGBWStrandTest was written before the gamma correction function was added to the library and implemented it themselves.

It also generated signed/unsigned comparison warnings,.

Both are now fixed,
Barometric pressure SPI example generated a warning correctly noting that a value was being leftshifted more bits than the size of it's datatype.

If we're gonna have examples. they might as well be correct, right?