Document overhead of Logic vs manual configuration
It can be summarized as  Don't worry about it on a 16 or 32k part. Be concerned 4./8k on 2-series and 2/4k on 0/1-series. (basically it's only an issue for the folks on 8-pin parts, and the small-flash 2-series parts.- Not nearly as bad as a lot of libraries, but much worse than some.