Fix assorted bugs.
Added missing standardized names to 0/1-series event generators.

Corrected a few tests.
Realized that TCBs are channel specific on 0/1-series. Which is a mondo bummer.

So we **STILL** don't have a way to make an input capture library that doesn't have to special case the 0/1-series tinies. However, I feel less bad about that with a way around any other special casing, and it being clearly microchip's dumb fault that it's like this.