Fix #625
Attaching interrupts to pins on PORTA with the default (all ports enabled) option would trigger BAD_ISR -> dirty reset -> reset  sequence... because I put the same test for a PORTx_PINS define that I do for the other ports. But I don't define PORTA pins because there is no part that would have been able to benefit from greater efficiency. Unfortunately I had forgotten that when I  tested for it to decide whether to define the ISR.

This is corrected and will be in the next release.