Fix #649
The previous write issue would manifest when writing a page of flash with a single word in it. This is the analogous case of reading a single "page" except that pages have no special meaning, The only limit is that we can only read 512 bytes at a time, (rep 255, ld *ptr++ where we don't send rep. It was that case that would fail