Readme updates, Wire loses requestFrom variants
These broke some libraries. Libraries would use a combination of data types that the compiles thought march equally well and reject it as abiguous (even though the numbers are all cast to 256 because maximum values passed for all three arguments that make sense in Arduinoland are already 256 or less, Since this allows math to be done with uint8_t's , it thus makes ssense to use that data type. All the other versions did was cast whwatever they were given to a uint8_t and call the base function. Which the compiler will do on it's own, using the "most appropriate" version of the function,. This change should hence have no change in behavior other than improved compatiblity and reduced binary size (very slightly) this will only impact sketches that were already broken because they were passing invalid values to the