fully add 2mhz non-tuned to all sections of boards.txt, adjust DIV
Not sure it is possible to correctly edit boards.txt but I think I found all 54 changes needed. Replaced the incorrect x16 div with a x10 div and adjusted the OSCCFG to use the 20mhz osc.

Also adjusted the OSCCFG for the tuned ones that were in there over and added a 2tuned to the 14-pin parts. Although less sure that will work, don't know if additional changes are needed for tuning. It looks like supported tuned speeds are not the same for all parts.