WARNING: CHECK MY ASSEMBLY CHANGE CAREFULLY
I don't know if there is a reason USART0_RXC_vect was supposed to reference &Serial1 but I changed it to 0 and it got rid of a compiler error

also the ifdef logic is wrong for the port register defines see my inserted and commented out #error("FUBAR") . Commenting out the part I commented out lets things compile. I suspect deleting these lines might be right but wanted @SpenceKonde to take a look at it before I did that.