add boards.txt audit/repair tool. use it to add missing non-tuned 2MHz options. Also fix a couple other bugs discovered in boards.txt
Point of change: 2Mhz was missing from many boards. But adding it manually is painful.

The generator that is checked in doesn't appear to work so I wrote a much simpler repair tool.

The way the tool works is you manually add the keys to a single board and then it will copy them to other boards. (you have to edit the script to tell it what to copy)

Also has some commented out prints that can help find missing things.